+++ src/ops/matmul/gemm.h
@@ -10,7 +10,7 @@ enum class Layout {
 };
 
 template <typename T>
-T defaultGEMMEpilogue(float acc, size_t im, size_t in) {
+T defaultGEMMEpilogue(float acc, int64_t im, int64_t in) {
     (void)im;
     (void)in;
     return llaisys::utils::cast<T>(acc);
@@ -40,40 +40,40 @@ void gemm_cpu_blocked_omp(
     constexpr int BK = 64;
 
 #pragma omp parallel for collapse(2) schedule(static)
-  for (size_t i0 = 0; i0 < M; i0 += BM) {
-    for (size_t j0 = 0; j0 < N; j0 += BN) {
+  for (int64_t i0 = 0; i0 < (int64_t)M; i0 += BM) {
+    for (int64_t j0 = 0; j0 < (int64_t)N; j0 += BN) {
 
-      size_t imax = std::min(i0 + BM, M);
-      size_t jmax = std::min(j0 + BN, N);
+      int64_t imax = std::min(i0 + BM, (int64_t)M);
+      int64_t jmax = std::min(j0 + BN, (int64_t)N);
 
       // tile-local float accumulator
       float acc_buf[BM][BN];
 
       // init
-      for (size_t ii = 0; ii < imax - i0; ++ii) {
-          for (size_t jj = 0; jj < jmax - j0; ++jj) {
+      for (int64_t ii = 0; ii < imax - i0; ++ii) {
+          for (int64_t jj = 0; jj < jmax - j0; ++jj) {
               acc_buf[ii][jj] = 0.0f;
           }
       }
 
-      for (size_t k0 = 0; k0 < K; k0 += BK) {
-          size_t kmax = std::min(k0 + BK, K);
+      for (int64_t k0 = 0; k0 < (int64_t)K; k0 += BK) {
+          int64_t kmax = std::min(k0 + BK, (int64_t)K);
 
-          for (size_t im = i0; im < imax; ++im) {
+          for (int64_t im = i0; im < imax; ++im) {
               const T* aptr = A + im * lda + k0;
               float*   acc_row = acc_buf[im - i0];
 
-              for (size_t k = k0; k < kmax; ++k) {
+              for (int64_t k = k0; k < kmax; ++k) {
                   float a = cast<float>(aptr[k - k0]);
 
                   if constexpr (LayoutB == Layout::RowMajor) {
                       const T* bptr = B + k * ldb + j0;
-                      for (size_t in = j0; in < jmax; ++in) {
+                      for (int64_t in = j0; in < jmax; ++in) {
                           acc_row[in - j0] +=
                               a * cast<float>(bptr[in - j0]);
                       }
                   } else {
-                      for (size_t in = j0; in < jmax; ++in) {
+                      for (int64_t in = j0; in < jmax; ++in) {
                           acc_row[in - j0] +=
                               a * cast<float>(B[in * ldb + k]);
                       }
@@ -83,11 +83,11 @@ void gemm_cpu_blocked_omp(
       }
 
       // epilogue + store
-      for (size_t im = i0; im < imax; ++im) {
+      for (int64_t im = i0; im < imax; ++im) {
           T* cptr = C + im * ldc + j0;
           float* acc_row = acc_buf[im - i0];
 
-          for (size_t in = j0; in < jmax; ++in) {
+          for (int64_t in = j0; in < jmax; ++in) {
               cptr[in - j0] =
                   epilogue(acc_row[in - j0], im, in);
           }
